Added target at 0, took care of some type conversion compiler warnings.

+++ code-experiments/src/coco_string.c
@@ -114,8 +114,8 @@ static long coco_strfind(const char *base, const char *seq) {
       }
       if (j == strlen_seq) {
         if (i > 1e9)
-          coco_error("coco_strfind(): strange values observed i=%lu, j=%lu, strlen(base)=%lu", i, j,
-              strlen(base));
+          coco_error("coco_strfind(): strange values observed i=%lu, j=%lu, strlen(base)=%lu",
+          		(unsigned long) i, (unsigned long) j, (unsigned long) strlen(base));
         return (long) i;
       }
     }
@@ -376,11 +376,13 @@ static size_t *coco_string_parse_ranges(const char *string,
       /* Make sure the boundaries are taken into account */
       if ((min > 0) && (num[0] < min)) {
         num[0] = min;
-        coco_warning("coco_string_parse_ranges(): '%s' ranges adjusted to be >= %lu", name, min);
+        coco_warning("coco_string_parse_ranges(): '%s' ranges adjusted to be >= %lu", name,
+        		(unsigned long) min);
       }
       if ((max > 0) && (num[1] > max)) {
         num[1] = max;
-        coco_warning("coco_string_parse_ranges(): '%s' ranges adjusted to be <= %lu", name, max);
+        coco_warning("coco_string_parse_ranges(): '%s' ranges adjusted to be <= %lu", name,
+        		(unsigned long) max);
       }
       if (num[0] > num[1]) {
         coco_warning("coco_string_parse_ranges(): '%s' ranges not within boundaries; some ranges ignored", name);
